requirements.txt
PyYAML
weasyprint==60.2
pydyf==0.10.0
